Keep installation recommendation tests in the CLI test module
